fix(stream): add missing webpack/rspack/rollup plugin entries
Extracts the streaming plugin hooks into a bundler-agnostic unplugin
factory (`packages/unhead/src/stream/unplugin.ts`) and exposes per-bundler
entry points:

- `unhead/stream/vite` → `createStreamingVitePlugin`
- `unhead/stream/webpack` → `createStreamingWebpackPlugin`
- `unhead/stream/rspack` → `createStreamingRspackPlugin`
- `unhead/stream/rollup` → `createStreamingRollupPlugin`

SSR detection is wired per bundler:
- vite: `ConfigEnv.isSsrBuild`, plus per-transform `environment.name` in
  vite v6+ dev
- webpack/rspack: `compiler.options.name === 'server'`

React/Solid/Svelte/Vue vite plugins now delegate to
`createStreamingVitePlugin`; the old standalone `createStreamingPlugin`
export is removed.
